clk: aspeed: ast2700: fix PLL rate calculation
The clk dump reported a wrong MPLL rate (e.g. 1010065408 instead of
1600000000), and every clock derived from it (CPU/AXI/PCLK) was off as
well.

The root cause is the bitfield layout of union ast2700_pll_reg. Mixing
uint16_t and uint8_t base types makes the struct occupy 8 bytes instead
of 4, so the uint32_t 'w' overlay is no longer aligned with the n/p
fields and the N field is decoded incorrectly. HPLL/DPLL happened to be
correct only because their N field is 0. For MPLL (reg 0x16300: m=768,
n=11) the broken union read n=1, giving mult=768/2=384 instead of
768/12=64, which then overflowed the u32 rate computation.

Make all bitfield members uint32_t so the struct packs into a single
32-bit word (sizeof == 4) and the fields map to the correct register
bits.

While at it, fix the M-PHY clock divider to use
SCU_CLKSEL1_MPHYCLK_DIV_MASK instead of the unrelated BCLK divider mask.

+++ arch/arm/include/asm/arch-aspeed/scu_ast2700.h
@@ -101,13 +101,13 @@
 union ast2700_pll_reg {
 	uint32_t w;
 	struct {
-		uint16_t m : 13;		/* bit[12:0]	*/
-		uint8_t n : 6;			/* bit[18:13]	*/
-		uint8_t p : 4;			/* bit[22:19]	*/
-		uint8_t off : 1;		/* bit[23]	*/
-		uint8_t bypass : 1;		/* bit[24]	*/
-		uint8_t reset : 1;		/* bit[25]	*/
-		uint8_t reserved : 6;		/* bit[31:26]	*/
+		uint32_t m : 13;		/* bit[12:0]	*/
+		uint32_t n : 6;			/* bit[18:13]	*/
+		uint32_t p : 4;			/* bit[22:19]	*/
+		uint32_t off : 1;		/* bit[23]	*/
+		uint32_t bypass : 1;		/* bit[24]	*/
+		uint32_t reset : 1;		/* bit[25]	*/
+		uint32_t reserved : 6;		/* bit[31:26]	*/
 	} b;
 };
